app/requirements.txt
Click==7.0
dnspython==1.15.0
eventlet==0.24.1
Flask==1.0.2
greenlet==0.4.15
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
monotonic==1.5
python-engineio==2.3.2
python-socketio==2.0.0
redis==2.10.6
six==1.11.0
Werkzeug==0.14.1
